requirements.txt
streamlit
langchain
python-dotenv
openai
langchain-community
langchain_openai
beautifulsoup4
pdfplumber
python-docx
pypandoc
sentence-transformers
chromadb